+++ fiddles/nativeScript/emulate.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-cd ~
-nohup bin/android-sdk-macosx/tools/emulator -netdelay none -netspeed full -avd 320480QVGA &