+++ scripts/bin/nativescript/_install.sh
@@ -13,6 +13,7 @@
 #  Revision History::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::|
 # ---------------------------------------------------------------------------------------------------|
 # Baseline Ver - CHANGELOG.MARKDOWN ~ 201605180420
+# 09/16/2016 - See CHANGELOG @ 201609160420
 # ---------------------------------------------------------------------------------------------------|
 
 this=$0;
@@ -35,17 +36,8 @@ function nativescriptInstall() {
     fi
 }
 
-function isNodeVer426Installed() {
-    if [[ ! -d "${NVM_DIR}/versions/node/v4.2.6" ]]
-    then
-        return 1;
-    else
-        return 0;
-    fi
-}
-
-function isNodeVer5Installed() {
-    if [[ ! -d "${NVM_DIR}/versions/node/v5.0.0" ]]
+function isNodeVerInstalled() {
+    if [[ ! -d "${NVM_DIR}/versions/node/v${__NODE_VERSION__}" ]]
     then
         return 1;
     else
@@ -67,32 +59,25 @@ function nvmInstall() {
     installed=$(isNvmInstalled;);
     if [[ "${installed}" == "false" ]]
     then
-        groupLog "${this}: nvm is not installed";
-        exit 1;
+        groupLog "${this}: nvm is not installed or NVM_DIR is not configured.";
+        groupLog "${this}: run the following command: ";
+        echo -e "\tnpm install -g nvm";
+        exit -1;
     else
-        if [[ $(isNodeVer5Installed;) -eq 1 ]]
+        if [[ $(isNodeVerInstalled;) -eq 1 ]]
         then
-            groupLog "${this}: node version 5.0 is not installed";
-            exit 1;
-        fi
-        if [[ $(isNodeVer426Installed;) -eq 1 ]]
-        then
-            groupLog "${this}: node version 4.2.6 is not installed.";
-            exit 1;
+            groupLog "${this}: node version ${__NODE_VERSION__} is not installed";
+            exit -1;
         fi
     fi
     . ~/.nvm/nvm.sh;
 }
 
-function nvmUseNodeVer5() {
-    groupLog "nvmUseNodeVer5";
-    nvm use 5.0 || exit $?;
+function nvmUseNodeVer() {
+    groupLog "nvmUseNodeVer";
+    nvm install ${__NODE_VERSION__} || exit $?;
 }
 
-function nvmUseNodeVer426() {
-    groupLog "nvmUseNodeVer426";
-    nvm use 4.2.6 || exit $?;
-}
 
 function adbInstall() {
     groupLog "adbInstall";
@@ -100,6 +85,6 @@ function adbInstall() {
     if [[ "${installed}" == "false" ]]
     then
         groupLog "${this}: adb is not installed";
-        exit 1;
+        exit -1;
     fi
 }